requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attr==0.3.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.33.11
botocore==1.33.11
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==38.0.4
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
deprecation==2.1.0
distro==1.8.0
dnspython==2.4.2
duckdb==0.8.1
duckduckgo-search==3.9.11
EbookLib==0.18
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.0
faiss-cpu==1.7.4
Faker==19.13.0
filelock==3.12.3
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
iopath==0.1.10
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lancedb==0.2.2
langchain==0.1.5
langchain-community==0.0.18
langchain-core==0.1.19
langdetect==1.0.9
langsmith==0.0.86
layoutparser==0.3.4
llvmlite==0.41.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.11
Markdown2docx==0.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.58.0
numexpr==2.8.5
numpy==1.25.2
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==1.7.0
openai-whisper==20230918
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pandasai==1.5.13
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
plantuml==0.3.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.6.3
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
pylance==0.6.5
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pypdfium2==4.20.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rapidfuzz==3.4.0
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
retry==0.9.2
rich==13.5.2
rpds-py==0.10.2
s3transfer==0.8.2
safetensors==0.3.3
scipy==1.11.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==1.4.49
stack-data==0.6.3
streamlit==1.31.0
streamlit-antd-components==0.2.4
streamlit-audiorec==0.1.3
streamlit-js-eval==0.1.5
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.3.3
timm==0.9.7
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.33.1
txtai==6.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.12
unstructured-inference==0.5.25
unstructured.pytesseract==0.3.12
urllib3==1.25.4
validators==0.22.0
wcwidth==0.2.8
wikipedia==1.4.0
xlrd==2.0.1
XlsxWriter==3.1.3
yarl==1.9.2
youtube-search-python==1.6.6
zipp==3.16.2
